base-files: fix debian build
Since the debian base-files don't have a /media, it can't be removed.
Hence test if the directory exists in the first place.

+++ meta-venus/recipes-core/base-files/base-files_%.bbappend
@@ -25,6 +25,8 @@ do_install_append() {
 	mkdir -p ${D}/opt
 	ln -s victronenergy ${D}/opt/color-control
 
-	rmdir ${D}/media
-	ln -s /run/media ${D}/media
+	if [ -d ${D}/media ]; then
+		rmdir ${D}/media
+		ln -s /run/media ${D}/media
+	fi
 }
